+++ packages/search-ui/src/Filter/ListFilter.tsx
@@ -21,6 +21,8 @@ const ListFilter = (props: Omit<ListFilterProps, 'type'>) => {
     itemRender,
     placeholder = '',
     format,
+    hideCount = false,
+    textTransform = 'normal-case',
   } = props;
 
   const filterContainerId = `list-${name}`;
@@ -41,7 +43,6 @@ const ListFilter = (props: Omit<ListFilterProps, 'type'>) => {
   const { disableDefaultStyles = false, customClassNames, currency, language } = useSearchUIContext();
   const theme = useTheme();
   const { t } = useTranslation('filter');
-
   const styles = getStylesObject(
     {
       innerList: [tw`flex items-center justify-between`],
@@ -114,9 +115,11 @@ const ListFilter = (props: Omit<ListFilterProps, 'type'>) => {
             css={styles.checkbox}
             disableDefaultStyles={disableDefaultStyles}
           >
-            {typeof itemRender === 'function' ? itemRender(label) : formatLabel(label, { format, currency, t })}
+            {typeof itemRender === 'function'
+              ? itemRender(label)
+              : formatLabel(label, { format, currency, textTransform, t })}
           </Control>
-          <span css={styles.count}>{count.toLocaleString(language)}</span>
+          {!hideCount && <span css={styles.count}>{count.toLocaleString(language)}</span>}
         </CoreBox>
       )),
     [JSON.stringify(items), itemRender, selected],