+++ packages/search-ui/src/Filter/SelectFilter.tsx
@@ -10,7 +10,15 @@ import { SelectFilterProps } from './types';
 import { formatLabel } from './utils';
 
 const SelectFilter = (props: Omit<SelectFilterProps, 'type'>) => {
-  const { name, title, sort = 'count', sortAscending = sort !== 'count', format } = props;
+  const {
+    name,
+    title,
+    sort = 'count',
+    sortAscending = sort !== 'count',
+    format,
+    textTransform = 'normal-case',
+    hideCount = false,
+  } = props;
   const { options, reset, setSelected, selected, multi } = useFilter(name, { sort, sortAscending });
   const { disableDefaultStyles = false, customClassNames, currency } = useSearchUIContext();
   const { t } = useTranslation('filter');
@@ -48,8 +56,8 @@ const SelectFilter = (props: Omit<SelectFilterProps, 'type'>) => {
           optionClassName={customClassNames.filter?.select?.option}
         >
           {options.map(({ value, label, count }) => (
-            <Option value={label} key={value} label={count.toLocaleString()}>
-              {formatLabel(label, { format, currency, t })}
+            <Option value={label} key={value} label={hideCount ? undefined : count.toLocaleString()}>
+              {formatLabel(label, { format, currency, textTransform, t })}
             </Option>
           ))}
         </Select>