+++ packages/search-ui/src/Filter/test/utils.test.ts
@@ -1,6 +1,7 @@
 import { FilterItem } from '@sajari/react-hooks';
 
-import { pinItems } from '../utils';
+import { TextTransform } from '../types';
+import { formatLabel, pinItems } from '../utils';
 
 describe('pinItems', () => {
   test.each([
@@ -77,4 +78,27 @@ describe('pinItems', () => {
   ])("pinItems(%o, %o, 'label')", (list: FilterItem[], pinned, expected) => {
     expect(pinItems(list, pinned, 'label')).toStrictEqual(expected);
   });
+});
+
+describe('formatLabel', () => {
+  const t = (s) => s;
+  test.each([
+    ['', 'normal-case', 'default', ''],
+    ['', 'uppercase', 'default', ''],
+    ['', 'capitalize', 'default', ''],
+    ['', 'capitalize-first-letter', 'default', ''],
+    ['all', 'capitalize-first-letter', 'default', 'All'],
+
+    ['text', 'normal-case', 'default', 'text'],
+    ['Car Electronics & GPS', 'uppercase', 'default', 'CAR ELECTRONICS & GPS'],
+    ['Car Electronics & GPS', 'lowercase', 'default', 'car electronics & gps'],
+    ['Car Electronics & GPS', 'capitalize', 'default', 'Car Electronics & GPS'],
+
+    ['> 200', 'normal-case', 'price', 'rangeOver'],
+  ])(
+    'formatLabel(%s, {textTransform:%s, format:%s})',
+    (label: string, textTransform: TextTransform, format: 'default' | 'price', expected) => {
+      expect(formatLabel(label, { textTransform, format, t })).toStrictEqual(expected);
+    },
+  );
 });